Fix: 트랜잭션 처리 개선 #182
- 데이터베이스에 조회한 뒤 업데이트하는 과정을 하나의 트랜잭션으로 처리하여 일관성 보장
